feat(M6-B): GPTQ PPL 측정 + fake_quant dtype 버그 수정
- demo.py에 W4A16 GPTQ 단계(3/6) 추가: _load_gptq_calib (wikitext-2 train 8×512)
- fake_quant_linear: weight_scale/input_scale을 w.dtype/x.dtype으로 캐스팅
→ float32 scale과 float16 텐서 혼용으로 F.linear dtype 불일치 방지
→ GPTQ에서 RuntimeError(c10::Half != float) 수정
→ W8A8 계열 fake quant가 float16 추론을 정확히 시뮬레이션하도록 수정

Qwen3-0.6B wikitext-2 PPL 측정 결과:
W4A16 GPTQ: 20.96 (+2.80) vs RTN: 25.89 (+7.73) → -4.93 개선
W8A8 static: 25.01 / SmoothQuant: 23.67 / W8A8 dynamic: 18.48

+++ mini_compressor/fake_quant_linear.py
@@ -65,8 +65,9 @@ class FakeQuantLinear(nn.Linear):
 
         qmax = 2 ** (spec.num_bits - 1) - 1
         qmin = -qmax if spec.symmetric else -(2 ** (spec.num_bits - 1))
-        s = self.weight_scale
-        zp = self.weight_zero_point if self.weight_zero_point is not None else torch.zeros_like(s)
+        s = self.weight_scale.to(w.dtype)
+        zp = (self.weight_zero_point.to(w.dtype)
+              if self.weight_zero_point is not None else torch.zeros_like(s))
 
         if spec.granularity == "per_channel":
             s = s.view(-1, *([1] * (w.dim() - 1)))
@@ -105,8 +106,9 @@ class FakeQuantLinear(nn.Linear):
         else:
             if self.input_scale is None:
                 return x
-            s = self.input_scale
-            zp = self.input_zero_point if self.input_zero_point is not None else torch.zeros_like(s)
+            s = self.input_scale.to(x.dtype)
+            zp = (self.input_zero_point.to(x.dtype)
+                  if self.input_zero_point is not None else torch.zeros_like(s))
 
         q = torch.clamp(torch.round(x / s + zp), qmin, qmax)
         return (q - zp) * s
